fix: updating xdrgen to only add fmtTest once to the go files
There are occasions when we will have imported fmt and not use it
depending on what we generate. To get around this being a problem we add
a line that sets the result of a fmt call to fmtTest. This way the
import is never unused. Unfortunatly we were adding fmtTest for each .x
file rather than in each .go file, this fixes that.

Breaks nothing