feat(plan): adding an execution plan to mazzaroth xdr
We need a data structure that encodes a series of transactions to
execute to a particular channel. The primary use case will be out of
band execution of transactions through a secure wallet.

One specific use case would be encoding this data as a QR code and
presenting it to a user from a website to allowing them to execute
transactions from a secure wallet app without exposing their private
accout key.

Breaks nothing