feat(account): some fixes for account information
The purpose of adding a StateRoot to readonly requests was to give the
developer an idea of the current status of state on the readonly node
they are interacting with. Unfortunately a state root that doesn't align
with a mined block is not very useful. Instead we'll return the most
recently mined block Hash and the number of transactions since it was
mined. This doesn't give a guarantee that they can look up the state at
the time of the request, but does give some information.

Other features include:
- Adding a general purpose AccountInfo lookup definition.
- Adding stateStatus to any request that doesn't hit the actual
consensus pool.

Breaks binaries using old mazzaroth xdr definitions.