fix: making the column type for array columns an array of size 1
Rust cannot have recursive types because their size cannot be known at
compile time.

This is fixed by making the column type for ArrayColumn a fixed array of 1.
This causes it to be wrapped in a Vec<Column> making it a dynamic type
which then has a known size at compile time.

Breaks nothing